requirements.txt
Django==4.2.7
djangorestframework==3.14.0
langchainplus-sdk==0.0.20
langchain==0.0.336
openai==0.28.1
chromadb==0.4.17
transformers==4.30.1
tiktoken==0.5.1
jsonlines==4.0.0
markdown==3.5.1
unstructured==0.10.30
requests==2.31.0
pyyaml==6.0.1
emoji==2.8.0
torch==2.1.0
rich==13.4.2
termcolor==2.3.0
python-dotenv==1.0.0
environs==9.5.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1